fix deleteMe

+++ app/Graphql/Mutations/DeleteMe.php
@@ -3,6 +3,8 @@
 namespace App\GraphQL\Mutations;
 
 use App\Models\User;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 use App\Http\Requests\UserRequest;
 use Illuminate\Support\Facades\Validator;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
@@ -15,9 +17,22 @@ final class DeleteMe
      */
     public function __invoke($_, array $args, GraphQLContext $context)
     {
-        $user = $context->request()->user();
-        $user->delete();
 
-        return $user ?? null;
+        $password = $args['input'];
+        $user = Auth::user();
+
+        if(Hash::check($password, $user->password)) {
+            $user->delete();
+            return [
+                'status' => 200,
+                'success' => true,
+            ];
+        }
+
+        return [
+            'status' => 401,
+            'success' => false,
+            'error' => 'The password doesn\'t match your actual password'
+        ];
     }
 }
